[sync] Migrate Olmo and Olmo2 to the Transformers modeling backend (#48100)
Ported the upstream OLMo migration: removed native OLMo, renamed the native OLMo2 implementation to OLMo3-only, routed OlmoForCausalLM and Olmo2ForCausalLM through the Transformers backend, and updated the supported-models docs. Verified edited Python files with py_compile and cleaned generated bytecode.

Upstream-vLLM: b83be00cddf03387bf740acfe6b9ca07ec1a3c08
